+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Changed the way of updating the tree
 * Welcome page with a quick start guide
 * Managing default ssh-user per kind via settings
+* Open in Browser for exposed ports
 
 ## [0.9.2] - 2025-02-27
 * TopoViewer Enhancements